+++ src/main/java/com/busanit/busan_subway_project/repo/LineRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LineRepo extends JpaRepository<Line, Long> {
-    // pk는 line_cd, Long 타입
-    Line findByLineCd(Long lineCd);
+public interface LineRepo extends JpaRepository<Line, Integer> {
+    // pk는 line_cd, Integer 타입
+    Line findByLineCd(int lineCd);
 }
